+++ projects/ontimize-web-ngx/src/lib/components/table/o-table.component.ts
@@ -229,7 +229,9 @@ export const DEFAULT_INPUTS_O_TABLE = [
   'showChartsOnDemandOption: show-charts-on-demand-option',
 
   // show-reset-width-option [yes|no|true|false]: show reset width menu option in the header menu
-  'showResetWidthOption: show-reset-width-option'
+  'showResetWidthOption: show-reset-width-option',
+
+  'disableSelectionFunction: disable-selection-function'
 ];
 
 export const DEFAULT_OUTPUTS_O_TABLE = [
@@ -244,6 +246,8 @@ const stickyFooterSelector = '.mat-footer-row .mat-table-sticky';
 const rowSelector = '.mat-row';
 const headerSelector = '.mat-header-row';
 const footerSelector = '.mat-footer-row';
+
+type DisableSelectionFunction = (item: any) => boolean;
 
 @Component({
   selector: 'o-table',
@@ -588,6 +592,7 @@ export class OTableComponent extends AbstractOServiceComponent<OTableComponentSt
   protected previousRendererData;
 
   quickFilterCallback: QuickFilterFunction;
+  disableSelectionFunction: DisableSelectionFunction;
 
   @ViewChild('tableBody', { static: false })
   protected tableBodyEl: ElementRef;
@@ -1972,10 +1977,20 @@ export class OTableComponent extends AbstractOServiceComponent<OTableComponentSt
     this.updateStateExpandedColumn();
   }
 
+  public getNumRowSelectedInCurrentData(): number {
+    return this.dataSource ? this.dataSource.renderedData.filter(x => !this.isDisableCheckbox(x) && this.isRowSelected(x)).length : 0;
+  }
+
   public isAllSelected(): boolean {
     const numSelected = this.selection.selected.length;
-    const numRows = this.dataSource ? this.dataSource.renderedData.length : undefined;
+    const numRows = this.getNumRowSelectedInCurrentData();
     return numSelected > 0 && numSelected === numRows;
+  }
+
+  public isIndeterminate(): boolean {
+    const numSelected = this.selection.selected.length;
+    const numRows = this.getNumRowSelectedInCurrentData();
+    return numSelected > 0 && numRows > 0 && numSelected !== numRows;
   }
 
   public masterToggle(event: MatCheckboxChange): void {
@@ -2008,6 +2023,9 @@ export class OTableComponent extends AbstractOServiceComponent<OTableComponentSt
   }
 
   public setSelected(item: any): void {
+    if (this.isDisableCheckbox(item)) {
+      return;
+    }
     if (Util.isDefined(item) && !this.isRowSelected(item)) {
       this.selection.select(item);
     }
@@ -3176,6 +3194,15 @@ export class OTableComponent extends AbstractOServiceComponent<OTableComponentSt
     this._oTableOptions.columns.forEach(c => {
       c.DOMWidth = this.getThWidthFromOColumn(c);
     });
+
+  }
+
+  public isDisableCheckbox(item: any): boolean {
+    let disable = false;
+    if (Util.isDefined(this.disableSelectionFunction)) {
+      return this.disableSelectionFunction(item);
+    }
+    return disable;
 
   }
 }